requirements.txt
typer[all]==0.12.3
requests>=2.31.0
rich==13.7.1
python-dotenv==1.0.1
